pyproject.toml
[tool.poetry]
name = "yt2script"
version = "0.1.0"
description = "YouTube to Transcript converter"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.3.3"
python-dotenv = "^1.0.0"
youtube-transcript-api = "^0.6.1"
requests = "^2.31.0"
gunicorn = "^21.2.0"
flask-cors = "^4.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
